+++ NEWS.md
@@ -8,6 +8,8 @@
   - fixed issue when package formula.tools was loaded due to overwriting of
     as.character.formula.
   - bug-fix: `regression(model, "y", "x", value="b")` now works as expected
+  - `index.estimate`, `index<-.estimate` methods for getting and setting
+    id/cluster
 
 # lava 1.9.1
   - `estimate.default`: removed `score.deriv`,`folds`, `R`, `null.sim` arguments