+++ NEWS.md
@@ -3,7 +3,10 @@
   - `merge.estimate` by default no longer sorts IC by id (old behaviour via new
     argument `sort=TRUE`)
   - influence function vignette updated.
-  - 
+  - `estimate.array` allows arguments `type="quantileN"` where N=1,..,9
+    corresponding to the type argument of `stats::quantile`. The `id` index is
+    now automatically derived from the row names unless explictly specified as
+    an argument.
 
 # lava 1.9.2
   - `estimate.default`: the `null`, `contrast`, `type`, and `var.adj` arguments